+++ src/store/types/filePaths.ts
@@ -1,0 +1,18 @@
+export type SpecificFiles = {
+	privacyExemplary: URL | string;
+	privacy: URL | string;
+	termsOfUseExemplary: URL | string;
+	termsOfUse: URL | string;
+	termsOfUseSchool: URL | string;
+	analogConsent: URL | string;
+};
+
+export type GlobalFiles = {
+	BeschreibungDerSchulCloud: URL | string;
+	TechnischerBericht2019: URL | string;
+	BroschuereSCimUnterricht1: URL | string;
+	BroschuereSCimUnterricht2: URL | string;
+	BroschuereSCundLernen4: URL | string;
+	SchulrechnerInDieSC2017: URL | string;
+	SCKonzeptPilotierung2017: URL | string;
+};